% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trie.R
\name{trie_find_subwords}
\alias{trie_find_subwords}
\title{Within a trie, find all possible words which can be created from the letters of the given word}
\usage{
trie_find_subwords(trie, word)
}
\arguments{
\item{trie}{root node of trie}

\item{word}{single character string}
}
\value{
character vector of words, or NULL if no matching words found
}
\description{
Within a trie, find all possible words which can be created from the letters of the given word
}
