% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trie.R
\name{trie_find_subwords_inner}
\alias{trie_find_subwords_inner}
\title{Find anagrams using all possible subsets of characters}
\usage{
trie_find_subwords_inner(node, chars)
}
\arguments{
\item{node}{starting node of tree}

\item{chars}{vector of single characters. Use '.' for a wildcard}
}
\value{
character vector of words. may contain duplicates
}
\description{
Find anagrams using all possible subsets of characters
}
