% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trie.R
\name{trie_insert}
\alias{trie_insert}
\title{Insert words into trie}
\usage{
trie_insert(trie, words)
}
\arguments{
\item{trie}{trie object}

\item{words}{character strings}
}
\description{
Insert words into trie
}
